feat: use new async queue system for faster sending (#41)
This is a rework of the `Kafee.Producer.AsyncWorker` module using some
of the things I learned in my work on stordco/data-streams-ex (mostly
using Task.async in a GenServer). It should hopefully be an easier read
now that the telemetry stuff is no longer leaking everywhere.

### Old and busted:
The GenServer runs a loop every 2 seconds. If there are messages in the
queue, it takes up to 100 messages from the queue, dispatches
`:brod.produce` in an async nature, waits for a response, and then takes
the messages out of the queue. This has a number of short comings being
that:

1) We possibly add 2 seconds of latency to message sending
2) We take a static batch size of messages to avoid hitting the message
size upper limit

### New hotness:
When the GenServer receives a message in the queue, it also adds a send
command to it's mailbox after 100ms. Once the send command runs, it
calls `Task.async` to run the `:brod.produce_sync` function. Once the
`Task.async` function finishes, it puts an ack command in the mailbox
along with how many messages it sent. The GenServer gets that message
and removes the messages out of the queue. This has a couple of
benefits:

1) We remove the 2 seconds of latency to maybe ~ 100ms (so we can batch
efficiently)
2) We batch messages based on size, so we can have larger batches for
smaller messages and vise versa
3) By using `:brod.produce_sync` in `Task.async`, we don't block mailbox
processing, but we can still use `:telemetry.span` to clean up all of
the telemetry reporting