fix: Return `:commit` since `:ack` does not commit in v2 (#84)
## Checklist

<!--
For each bullet, ensure your pr meets the criteria and write a note
explaining how this PR relates. Mark them as complete as they are done.
All top-level checkboxes should be checked regardless of their relevance
to the pr with a note explaining whether they are relevant or not.
-->

- [x] Code conforms to the [Elixir
Styleguide](https://github.com/christopheradams/elixir_style_guide)

## Problem

<!--
What is the problem you're solving or feature you're implementing? Link
to any Jira tickets or previous discussions of the issue.
-->

When using the `brod_group_subscriber_v2` our `handle_message/3`
callback was returning `{:ok, :ack, state}`. This was the old way to
commit messages; in v2, it only acks and doesn't commit based on this
comment:

https://github.com/kafka4beam/brod/blob/master/src/brod_group_subscriber_worker.erl#L85

## Details

<!--
Include a brief overview of the technical process you took (or are going
to take!) to get from the problem to the solution.
-->

Changed to commit