fix: UPDATE: kafee async worker to log only the headers + topic for large message error instead of the payload as it can go over the datadog limit (#108)
## Related Ticket(s)
SIGNAL-7183
<!--
Enter the Jira issue below in the following format: PROJECT-##
-->

## Checklist

<!--
For each bullet, ensure your pr meets the criteria and write a note
explaining how this PR relates. Mark them as complete as they are done.
All top-level checkboxes should be checked regardless of their relevance
to the pr with a note explaining whether they are relevant or not.
-->

- [x] Code conforms to the [Elixir
Styleguide](https://github.com/christopheradams/elixir_style_guide)

## Problem

https://stord.atlassian.net/wiki/spaces/PD/pages/3528982601/WMS+integration+bridge+receipt+confirmations+triaging

We had an issue where large messages didn't even log, it silently
omitted.
Suspicion is the message being too large for Datadog.

## Details

We need at least something to help out with tracing which message or
data the missed message was for.
Therefore, instead of the actual body of the message, we will log the
metadata + topic of the message.