chore: add consumer lag utility for brod (#131)
## Related Ticket(s)
SIGNAL-7544
<!--
Enter the Jira issue below in the following format: PROJECT-##
-->

## Checklist

<!--
For each bullet, ensure your pr meets the criteria and write a note
explaining how this PR relates. Mark them as complete as they are done.
All top-level checkboxes should be checked regardless of their relevance
to the pr with a note explaining whether they are relevant or not.
-->

- [x] Code conforms to the [Elixir
Styleguide](https://github.com/christopheradams/elixir_style_guide)

## Problem

<!--
What is the problem you're solving or feature you're implementing? Link
to any Jira tickets or previous discussions of the issue.
-->
There needs a way to programmatically retrieve consumer lag data, per
partition per consumer group.


## Details
The module is a collection of utility functions which are light wrappers
around `:brod` functions.
The logic has been tested during consumer lag spikes in production to
show the differences.

Record of the tests are here in this google
[doc](https://docs.google.com/document/d/1cS_GVfu7htdYaFNPxHd0rvXPRVDfRmfYbjfLVbb395I/edit?tab=t.0).
<!--
Include a brief overview of the technical process you took (or are going
to take!) to get from the problem to the solution.
-->

---------

Co-authored-by: Matt Sutkowski <msutkowski@gmail.com>