fix: test to be more accurate (#133)
## Related Ticket(s)
[SIGNAL-7636](https://stord.atlassian.net/browse/SIGNAL-7636)
<!--
Enter the Jira issue below in the following format: PROJECT-##
-->

## Checklist

<!--
For each bullet, ensure your pr meets the criteria and write a note
explaining how this PR relates. Mark them as complete as they are done.
All top-level checkboxes should be checked regardless of their relevance
to the pr with a note explaining whether they are relevant or not.
-->

- [x] Code conforms to the [Elixir
Styleguide](https://github.com/christopheradams/elixir_style_guide)

## Problem
There is a unit test that needs to be adjusted to ensure expected map
struct is compared in equality, not pattern matched.

<!--
What is the problem you're solving or feature you're implementing? Link
to any Jira tickets or previous discussions of the issue.
-->

## Details
The test is adjusted.
<!--
Include a brief overview of the technical process you took (or are going
to take!) to get from the problem to the solution.
-->


[SIGNAL-7636]:
https://stord.atlassian.net/browse/SIGNAL-7636?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ